Add long status option to `battery`
The long version uses `ioreg` rather than `pmset`. This can be expanded
in the future to include an indicator of battery health.

Implementing this requires a somewhat broken approach since there isn't
currently a working argument array for functions. The latest version of
bash boilerplace includes such an array and will need to be updated in a
later commit.