Add `help`/`desc` support for colons and hyphens in command names.
Bash allows ":" and "-" in function names, but they are not allowed in
identifiers for variables. In order to store the description in a variable
that matches the function name, this functionality escapes the function name
in order to make it suitable for use in a variable identifier.

These changes are copied directly from the `bask` framework:
https://github.com/alphabetum/bask