Use hyphens in program-specific commands with two-word names.
Functions with two word names are expected to use hyphens rather than
underscores to stick with common conventions and make it easier to type
them. In addition to updating function names, also update the `desc`
blocks and function usage info in Readme.

Some functions with two-word names are named directly after the original
utilities included with OS X. In these cases, the convention used for
the original utility name is used for the command name.

Also use this opportunity to correct ome mistakes in the `login-items` Readme
usage info.